0.2.6.0 -fps fix mit anzeige

+++ src/main/java/theCreepyProperty/entity/Entity.java
@@ -11,7 +11,7 @@ public class Entity {
     public int keys_eingesammelt = 0;
     public Image up1, up2, up3, up4, down1, down2, down3, down4, left1, left2, left3, left4, right1, right2, right3, right4;
     public String direction;
-    public int sprite_counter = 0;
+    public double sprite_counter = 0;
     public int sprite_num = 1;
     public Rectangle solid_player_aria;
     public double solid_aria_default_x, solid_aria_default_y;
